refactor(desktop): isolate app server client pool
